+++ build-logic/src/main/kotlin/buildLogic/gradleRunCommandLine.kt
@@ -1,0 +1,9 @@
+package buildLogic
+
+import org.apache.tools.ant.taskdefs.condition.Os
+
+fun gradleRunCommandLine(vararg arguments: String) = if (Os.isFamily(Os.FAMILY_WINDOWS)) {
+    listOf("cmd", "/c", "gradlew.bat", *arguments)
+} else {
+    listOf("./gradlew", *arguments)
+}